Add cache_ok=True class var to UUIDType
Since SQLAlchemy 1.4.14, following warning was emitted:

"""
/sqlalchemy/sql/type_api.py:1104: SAWarning: TypeDecorator UUIDType(binary=False) will not produce a cache key
because the `cache_ok` flag is not set to True.  Set this flag to True if this type object's state is safe to
use in a cache key, or False to disable this warning
"""

To prevent this, `cache_ok = True` is introduced as a class variable.
Both `binary` and `native` parameters in `__init__` method are `bool`,
so safely hashable, but the `**kwargs` may not be. Since `kwargs` are
not used anywhere in the initializer, they are deleted as a safety
measure.