Update color.py for colour-science compatibility
This will allow SQLAlchemy-Utils to be compatible with `colour-science`
Currently, this import makes this package incompatible with `colour-science` due to conflicting `colour` namespace. Since the `colour` package is already optional, this will allow `sqlalchemy-utils` to be used with `colour-science` without conflict.
see [here](https://github.com/kvesteri/sqlalchemy-utils/issues/636) for the associated issue.