Address pytest warnings that test functions should return `None`
This also fixes a mistake in the test suite for Path fields:
If the path is empty, then due to the behavior of `path.split()`,
its length is zero.

It might be preferable to make an empty path return a length of 0,
but that means it also might be preferable to change the behavior
of the `parts` property, which is not in scope for this work.