add schema support for view create/drop
previous functionality did not support create/drop of views within a
schema. This causes issues when a users is expecting the view to exist
within the schema specified by the metadata object - but it is instead
created in the default schema regardless.

adding in fuctionality to support creation of views within a schema as
provided by the metadata object. Defaults to None if no schema is
present.

adding in functionality to specify schema when dropping view as provided
by metadata object

modified:   sqlalchemy_utils/view.py